rename Blacklight::SearchContextComponent to Blacklight::SearchContext::ServerItemPaginationComponent
